stdiokey-python/pyproject.toml
[project]
name = "stdiokey-python"
version = "0.4.0"
description = "System Utilities with API Key MCP Stdio transport in Python"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.2.1",
    "psutil>=6.1.1",
    "google-api-python-client>=2.160.0",
    "google-auth-oauthlib>=1.2.1",
    "pydantic>=2.10.6",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
